Python Work
I started a new directory for python work as I would like to refresh my
working knowledge of Python so I can use it as a fun/fast idea
prototyping language. Of course I had to create the customary "hello,
world" program to get things started off with a bang.

08/29/2020