Class Review
I build a simple car project to test out my knowledge of how to create
classes and child classes in python. This was very easy and it felt good
to know that I retained much of what I studied a few years ago. I can
tell that my ability to jump between languages and frameworks has
improved in the past 6 months. Now I'm going to work with files and
exceptions. I want to start working with data so this is a natural
progression for my study. By the end of the day my goal is to be writing
a few programs and visualizing my work.

08/31/2020